backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1
PyJWT==2.8.0
